fix cross correlation output in signed 32bits

+++ stm32_exercise/Core/Inc/api_utils.h
@@ -58,7 +58,7 @@ void invertir (uint16_t * vector, uint32_t longitud);
 /**
  * Cross correlation calculation for 2 vectors
  */
-void corr (int16_t *vectorX, int16_t * vectorY, int16_t * vectorCorr, uint32_t longitud);
+void corr (int16_t *vectorX, int16_t * vectorY, int32_t * vectorCorr, uint32_t longitud);
 
 
 #endif /* INC_API_UTILS_H_ */
